added auth composable for better readability
